skills/gemini-video/runtime/pyproject.toml
[project]
name = "gemini-video"
version = "0.3.0"
description = "Generate videos using Google Gemini Veo API"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
]

[dependency-groups]
dev = [
    "pillow>=12.1.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.pyright]
pythonVersion = "3.10"
include = ["."]
typeCheckingMode = "standard"
